Update : 인증 재확인

+++ src/main/java/com/example/peeppo/domain/bid/service/BidService.java
@@ -7,6 +7,7 @@ import com.example.peeppo.domain.bid.dto.BidListResponseDto;
 import com.example.peeppo.domain.bid.dto.ChoiceRequestDto;
 import com.example.peeppo.domain.bid.entity.Bid;
 import com.example.peeppo.domain.bid.entity.Choice;
+import com.example.peeppo.domain.bid.enums.GoodsStatus;
 import com.example.peeppo.domain.bid.repository.BidRepository;
 import com.example.peeppo.domain.bid.repository.ChoiceBidRepository;
 import com.example.peeppo.domain.bid.repository.QueryRepository;
@@ -42,90 +43,56 @@ public class BidService {
     private final QueryRepository queryRepository;
     private final ChoiceBidRepository choiceBidRepository;
 
-    public ResponseEntity<ResponseDto> bidding(User user, Long auctionId, BidGoodsListRequestDto bidGoodsListRequestDto) {
+    public ResponseEntity<ResponseDto> bidding(User user, Long auctionId, BidGoodsListRequestDto bidGoodsListRequestDto) throws IllegalAccessException {
 
         Auction auction = getAuction(auctionId);
-
-//        Long goodsId = bidRequestDto.getGoodsId();
-//        Goods goods = getGoods(goodsId);
-//        String goodsImg = String.valueOf(imageRepository.findFirstByGoodsGoodsIdOrderByCreatedAtAsc(goodsId));
-
-//        if(!goods.isDeleted()){
-////            ResponseDto response = new ResponseDto("삭제된 물품입니다.", HttpStatus.BAD_REQUEST.value(), "BAD_REQUEST");
-////            return ResponseEntity.status(HttpStatus.BAD_REQUEST.value()).body(response);
-//            throw new IllegalArgumentException();
-//        }
+        List<Bid> List = new ArrayList<>();
 
         //경매 진행 여부
-//        validateAuctionPeriod(auction.getAuctionStatus());
+//        if (auction.getGoods().getGoodsStatus().equals(GoodsStatus.ONAUCTION.getStatus())) {
+            if (!auction.getUser().getUserId().equals(user.getUserId())) {
+                for (Long goodsId : bidGoodsListRequestDto.getGoodsId()) {
+                    Goods goods = getGoods(goodsId);
+                    String goodsImg = String.valueOf(imageRepository.findFirstByGoodsGoodsIdOrderByCreatedAtAsc(goodsId));
+
+                    if (goods.isDeleted()) {
+                        continue;                           //여기도 고민
+                    }
+                    if(!goods.getUser().equals(user)){
+                        throw new IllegalAccessException();
+                    }
+
+//                    if(goods.getGoodsStatus().equals(GoodsStatus.ONSALE.getStatus())){
+                        //시작가보다 낮을 경우
+//                    if (auction.getLowPrice() > ) {
 //
-//        //경매 등록한 사람이랑 같은지 확인
-//        if (auction.getUser().getId().equals(userId)) {
-//            throw new IllegalArgumentException();
+//                    }
+                        List.add(new Bid(user, auction, goods, goodsImg));
+//                    }
+//                    else {
+//                        throw new IllegalAccessException(); //어떤 예외처리를 할까
+//                    }
+                }
+            }
+            else {
+                throw new IllegalAccessException(); //어떤 예외처리를 할까
+            }
 //        }
-//
+//        else {
+//            throw new IllegalAccessException(); //어떤 예외처리를 할까
+//        }
+
 //        //시작가보다 낮을 경우
 //        if(auction.getGoods().getGoodsRating() > goods.getGoodsRating()){
 //            return;
 //        }
 //        else {
 
-        List<Bid> List = new ArrayList<>();
-
-        for (Long goodsId : bidGoodsListRequestDto.getGoodsId()) {
-            Goods goods = getGoods(goodsId);
-            String goodsImg = String.valueOf(imageRepository.findFirstByGoodsGoodsIdOrderByCreatedAtAsc(goodsId));
-
-            if (goods.isDeleted()) {
-                continue;
-            }
-            List.add(new Bid(user, auction, goods, goodsImg));
-        }
-
-//        Bid bid = new Bid(user, auction, goods, goodsImg);
         bidRepository.saveAll(List);
 
-        ResponseDto response = new ResponseDto("등록 완료", HttpStatus.OK.value(), "OK");
+        ResponseDto response = new ResponseDto("입찰이 완료되었습니다.", HttpStatus.OK.value(), "OK");
         return ResponseEntity.status(HttpStatus.OK.value()).body(response);
 //        }
-    }
-
-
-    //물품 수정
-    public ResponseEntity<ResponseDto> updateBid(User user, Long auctionId, Long bidId, BidGoodsListRequestDto bidGoodsListRequestDto) {
-        Auction auction = getAuction(auctionId);
-        Bid bid = getBid(bidId);
-
-        for (Long goodsId : bidGoodsListRequestDto.getGoodsId()) {
-            Goods goods = goodsRepository.findById(goodsId).orElseThrow(() -> new IllegalArgumentException());
-            String goodsImg = String.valueOf(imageRepository.findFirstByGoodsGoodsIdOrderByCreatedAtAsc(goodsId));
-
-            if (goods.isDeleted()) {
-                continue; //이미 삭제된 goodsId라는거 표시해주기
-            }
-            bid.update(user, auction, goods, goodsImg);
-        }
-
-        bidRepository.save(bid);
-
-        ResponseDto response = new ResponseDto("재등록 완료", HttpStatus.OK.value(), "OK");
-        return ResponseEntity.status(HttpStatus.OK.value()).body(response);
-    }
-
-    //해당 유저, 해당 옥션에 전부를 취소
-    @Transactional
-    public ResponseEntity<ResponseDto> deleteBid(User user, Long auctionId) {
-        Auction auction = getAuction(auctionId);
-        Long userId = user.getUserId();
-
-        List<Bid> bidList = queryRepository.findBid(user.getUserId(), auctionId);
-
-        for (Bid bid : bidList) {
-            bidRepository.delete(bid);
-        }
-
-        ResponseDto response = new ResponseDto("삭제 완료", HttpStatus.OK.value(), "OK");
-        return ResponseEntity.status(HttpStatus.OK.value()).body(response);
     }
 
     public ResponseEntity<Page<BidListResponseDto>> BidList(Long auctionId, int page, int size, String sortBy, boolean isAsc) {
@@ -146,52 +113,39 @@ public class BidService {
     }
 
     //경매자가 선택, try문을 하는게 깔끔하려나?
-    public ResponseEntity<ResponseDto> choiceBids(User user, Long auctionId, ChoiceRequestDto choiceRequestDto) {
+    public ResponseEntity<ResponseDto> choiceBids(User user, Long auctionId, ChoiceRequestDto choiceRequestDto) throws IllegalAccessException {
         Auction auction = getAuction(auctionId);
         List<Choice> bidsList = new ArrayList<>();
 
-//        if (auction.getUser().getUserId() == user.getUserId()) {
-//            bidRepository.saveAll(biddingList(bidsList, bidGoodsListRequestDto, user, auction));
-//
-//            ResponseDto response = new ResponseDto("등록 완료", HttpStatus.OK.value(), "OK");
-//        }
-//        else{
-//            ResponseDto response = new ResponseDto("해당 경매 물품의 판매자가 아닙니다.", HttpStatus.BAD_REQUEST.value(), "BAD_REQUEST");
-//        }
-
-//        if (auction.getUser().getUserId() == user.getUserId()) {
-        for (Long bidId : choiceRequestDto.getbidId()) {
-            Bid bid = getBid(bidId);
-            bidsList.add(new Choice(bid, auction));
+        if (auction.getUser().getUserId().equals(user.getUserId())) {
+            getBiddingList(choiceRequestDto, auction, bidsList);
+        }
+        else {
+            throw new IllegalAccessException();
         }
 
-        choiceBidRepository.saveAll(bidsList);
-//          }
-
-        ResponseDto response = new ResponseDto("선택 완료", HttpStatus.OK.value(), "OK");
+        ResponseDto response = new ResponseDto("선택이 완료되었습니다.", HttpStatus.OK.value(), "OK");
         return ResponseEntity.status(HttpStatus.OK.value()).body(response);
     }
 
     //경매자가 선택 바꾸는 기능
-    public ResponseEntity<ResponseDto> choiceUpdateBids(User user, Long auctionId, ChoiceRequestDto choiceRequestDto) {
+    public ResponseEntity<ResponseDto> choiceUpdateBids(User user, Long auctionId, ChoiceRequestDto choiceRequestDto) throws IllegalAccessException {
         Auction auction = getAuction(auctionId);
         List<Choice> bidsList = new ArrayList<>();
         List<Choice> findAllChoice = queryRepository.findChoice(auctionId);
 
-//        if (auction.getUser().getUserId() == user.getUserId()) {
-        for (Choice choice : findAllChoice) {
-            choiceBidRepository.delete(choice);
+        if (auction.getUser().getUserId().equals(user.getUserId())) {
+            for (Choice choice : findAllChoice) {
+                choiceBidRepository.delete(choice);
+            }
+
+            getBiddingList(choiceRequestDto, auction, bidsList);
+        }
+        else {
+            throw new IllegalAccessException();
         }
 
-        for (Long bidId : choiceRequestDto.getbidId()) {
-            Bid bid = getBid(bidId);
-            bidsList.add(new Choice(bid, auction));
-        }
-
-        choiceBidRepository.saveAll(bidsList);
-//        }
-
-        ResponseDto response = new ResponseDto("선택 수정 완료", HttpStatus.OK.value(), "OK");
+        ResponseDto response = new ResponseDto("재선택 되었습니다.", HttpStatus.OK.value(), "OK");
         return ResponseEntity.status(HttpStatus.OK.value()).body(response);
     }
 
@@ -211,6 +165,18 @@ public class BidService {
         return bidList;
     }
 
+    private void getBiddingList(ChoiceRequestDto choiceRequestDto, Auction auction, List<Choice> bidsList) throws IllegalAccessException {
+        for (Long bidId : choiceRequestDto.getbidId()) {
+            Bid bid = getBid(bidId);
+
+            if(!auction.getAuctionId().equals(bid.getAuction().getAuctionId())){
+                throw new IllegalAccessException();
+            }
+            bidsList.add(new Choice(bid, auction));
+        }
+
+        choiceBidRepository.saveAll(bidsList);
+    }
 
     private Bid getBid(Long bidId) {
         return bidRepository.findById(bidId).orElseThrow(
